Add the spinning animation.
Click on a cube and it will start spinning.
Then each of the cubes behind it will start spinning in turn.
(Because you can only click on the ones in the front.)
The transform origin is now the center of cube in all three dimensions.

Add help and README.

Adjust the perspective so its not so extreme.